interview_task/pyproject.toml
[tool.poetry]
name = "interview-task"
version = "0.1.0"
description = ""
authors = ["Anastasia Vakhrusheva <aaselezneva9@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.8.1"
attrs = "24.2.0"
black = "24.8.0"
click = "8.1.7"
django = "5.1"
django-filter = "24.3"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.27.2"
environs = "11.0.0"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
marshmallow = "3.22.0"
mypy-extensions = "1.0.0"
packaging = "24.1"
pathspec = "0.12.1"
platformdirs = "4.2.2"
psycopg2-binary = "2.9.9"
pyjwt = "2.9.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
referencing = "0.35.1"
rpds-py = "0.20.0"
sqlparse = "0.5.1"
uritemplate = "4.1.1"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.7.4"
cffi = "1.17.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
coverage = "7.6.1"
crashtest = "0.4.1"
distlib = "0.3.8"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
idna = "3.7"
importlib-metadata = "8.4.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
more-itertools = "10.4.0"
msgpack = "1.0.8"
pexpect = "4.9.0"
pkginfo = "1.11.1"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pycparser = "2.22"
pyproject-hooks = "1.1.0"
rapidfuzz = "3.9.6"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
tomlkit = "0.13.2"
trove-classifiers = "2024.7.2"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
xattr = "1.1.0"
zipp = "3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
